pyproject.toml
[project]
name = "faltu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "alembic==1.13.1",
    "dependency-injector==4.41.0",
    "fastapi==0.111.0",
    "groq==0.9.0",
    "httpx==0.27.0",
    "loguru==0.7.2",
    "numpy==1.26.4",
    "opencv-python-headless==4.9.0.80",
    "pillow==10.3.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.1",
    "pydantic-settings==2.3.0",
    "pytesseract==0.3.10",
    "pytest==8.2.1",
    "pytest-asyncio==0.23.7",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "requests==2.32.3",
    "sqlalchemy==2.0.30",
    "streamlit==1.35.0",
    "uvicorn==0.30.0",
]
